Find Dafny you’re special

+++ codegen/smithy-dafny-codegen/src/main/java/software/amazon/polymorph/CodegenEngine.java
@@ -133,10 +133,13 @@ public class CodegenEngine {
         for (final TargetLanguage lang : targetLangOutputDirs.keySet()) {
             final Path outputDir = targetLangOutputDirs.get(lang).toAbsolutePath().normalize();
 
-            // Clear out all contents first to make sure if we didn't generate it,
-            // it doesn't show up.
-            software.amazon.smithy.utils.IoUtils.rmdir(outputDir);
-            outputDir.toFile().mkdirs();
+            // Clear out all contents first to make sure if we didn't intend to generate it,
+            // it doesn't show up as generated code.
+            // We can't do this for Dafny because we like to generate into the Smithy Model directory :P
+            if (lang != TargetLanguage.DAFNY) {
+                software.amazon.smithy.utils.IoUtils.rmdir(outputDir);
+                outputDir.toFile().mkdirs();
+            }
 
             switch (lang) {
                 case DAFNY -> generateDafny(outputDir);
